Add Hanami new, excluding .env files
I'm not sure how Hanami advises handling secrets, but the dev and test `.env` files both contain a session secret.

I suppose you could easily argue that these environments will never be exposed to anyone and therefore that this is safe.

I went back and forth on this but ultimately decided not to commit these files for now.