Add index.html
For starters we're just going to do the first paragraph of the prologue of Unamuno's seminal work 'Niebla'. This isn't simple by any stretch, but it is short and interesting.

I have pulled over a bunch of functionality from the POC of the tibetan translation tool this library will ultimately be consumed by.