Make Cypress at home
Commit all the crap (probably mostly unwanted) that came with it, but in an appropriate corner of the app that doesn't make Cypress a first-class citizen.

Make sure videos and screenshots won't be committed.

Honestly I've been wrong many times in my life, but so far Cypress seems very much in favor of making itself the star of the show and doing things for you in an opinionated and magic way.

I'm leaning towards building a UI client for browser acceptance testing that wraps Cypress functionality for now, writing a few more tests and/or doing so more research, and then likely ripping Cypress out to replace with Nightwatch, Jest with Puppeteer, or something else.

One thing I'm 100% clear I want for the acceptance tests:

Dedicated clients written in strongly-typed JavaScript

Adds some acceptance test musings to the planning directory.